requirements.txt
opencv-python
numpy
matplotlib
numpy-stl
python-telegram-bot
